Add VPC awareness to network data source (#2547)
* Add VPC awareness to network data source

Network data source uses network name to identify a network.
However, this is problematic with VPC subnets, which appear as a network in a
similar way to host PGs and DVPGs. As subnet names aern't unique across VPCs,
there is no way to distingish these other than using the full path.

To address this, added the VPC id which is used in the same way as DVS id can
be specified, to scope the network search to the VPC.

As govmomi has no VPC capabilities within the network scope, parsed the object
path to conclude the VPC id.

Signed-off-by: Kobi Samoray <kobi.samoray@broadcom.com>

* Update vsphere/internal/helper/network/network_helper.go

Co-authored-by: Stoyan Zhelyazkov <156204153+spacegospod@users.noreply.github.com>

---------

Signed-off-by: Kobi Samoray <kobi.samoray@broadcom.com>
Co-authored-by: Stoyan Zhelyazkov <156204153+spacegospod@users.noreply.github.com>