fix: check for MissingPortGroupReferenceError when reading a VM network subresource
The govmomi SOAP call for looking up a dvportgroup can fail in two ways: with a SOAP fault (which gets translated to a Go error containing the substring "The object or item referred to could not be found"), or with a successful response containing an empty return value. In the latter case, the response is translated to a MissingPortGroupReferenceError Go error by the dvportgroup.FromKey method. In both cases, we want to ignore the error and consider that the VM resource has no associated network. This allows cloning a VM from a template associated with a network the user does not have access to.

Closes: #2586

Signed-off-by: Joan Sala Isern <jsiwrk@gmail.com>